Fix a bug and extended functionality.
If a signal name was longer than the state name the signal names were
being over written with part of the arrow in the sequence diagram.
Changed how the signal name was placed within the arrow to fix this
issue.

Added pre-filtering to the sequence to get right of snoop trace/spy
information before parsing the resulting trace into a sequence diagram.